==> Call Center Anomaly Detection/1-EDA.ipynb <==
# IMPORTING LIBRARIES




import pandas as pd
import numpy as np
import seaborn as sns
from datetime import datetime

from tensorflow import keras
from tensorflow.keras import layers


from google.colab import drive
drive.mount('/content/drive')

df_lokasyon = pd.read_csv("/content/drive/MyDrive/new_csv.csv")

replaceDict={'Ý':'İ','Ð':'Ğ','Þ':'Ş','þ':'ş','ý':'ı','ð':'ğ'}

for x in replaceDict:
    df_lokasyon=df_lokasyon.replace(x,replaceDict[x],regex=True)

# EXAMINATION OF DATA




df_lokasyon

df_lokasyon.describe()

df_lokasyon.info()




## MISSING VALUES

sns.heatmap(df_lokasyon.isnull(),yticklabels=False,cbar=False,cmap='viridis')

df_lokasyon["X_IBB_NEIGHBOURHOOD"].isna().sum()/df_lokasyon["X_IBB_NEIGHBOURHOOD"].shape[0]

df_lokasyon.isna().sum()

## ARRANGEMENT OF COLUMN NAMES











df_lokasyon = df_lokasyon.rename(columns={'ORG_TICKET_ID': 'BASVURU_KODU', 'OPEN_DT': 'ACILIS_TARIHI', 'ACT_CLOSE_DT': 'KAPANIS_TARIHI', 'TICKET_DESC': 'BASVURU_TANIMI','X_IBB_RELATED_DIV': 'ILGILI_BIRIM', 'SR_AREA': 'KONU', 'SR_SUB_AREA': 'ALT_KONU', 'SR_SUBTYPE_CD': 'KANAL','SR_STAT_ID': 'DURUM', 'X_IBB_COUNTY': 'ILCE', 'X_IBB_NEIGHBOURHOOD': 'MAHALLE'})

df_lokasyon.head()

## EXAMINATION OF EACH COLUMN

İLÇE BAZINDA İNCELEME YAPTIĞIMIZDA BURADA İSTANBUL'DA YER ALMAYAN İLÇELERDEN DE ÇEŞİTLİ ÇAĞRILARIN GELDİĞİNİ TESPİT ETTİK.** İSTANBUL'DA 39 ADET İLÇE** BULUNMASINA RAĞMEN İLÇE SAYISI BAZINDA İNCELENDİĞİNDE **769 FARKLI İLÇEDEN ** GELEN ÇAĞRILAR MEVCUT. YAPACAĞIMIZ ÇALIŞMA İSTANBUL'U KAPSADIĞI İÇİN İSTANBUL DIŞI İLÇE VERİLERİNİ DF_LOKASYON VERİSİNDEN ATTIK.

df_lokasyon['ILCE'].unique().shape

df_lokasyon['ILCE'].value_counts()[:]

İSTANBUL'DA 39 İLÇE YER ALIYOR, BİZ İLK 40 İLÇEYE BAKTIK VE 40. İLÇEDEN GELEN ÇAĞRI SAYISINI BAZ ALARAK İSTANBUL'DA OLMAYAN İLÇELERİ DÜŞÜRDÜK







df_lokasyon['ILCE'].value_counts()[:50]

counts = df_lokasyon['ILCE'].value_counts()
df_loc = df_lokasyon[~df_lokasyon['ILCE'].isin(counts[counts < 2345].index)]

df_loc['BASVURU_KODU'].describe()

df_loc['ACILIS_TARIHI'].describe()

df_loc['KAPANIS_TARIHI'].describe()

df_loc['BASVURU_TANIMI'].describe()

df_loc['ILGILI_BIRIM'].describe()

df_loc['KONU'].describe()

df_loc['KONU'].value_counts().head(10)

df_loc['ALT_KONU'].describe()

df_loc['ALT_KONU'].value_counts().head(10)

df_loc['KANAL'].describe()

df_loc['DURUM'].describe()

df_loc['ILCE'].describe()

df_loc['MAHALLE'].describe()

335 TANE EŞSİZ KONU DEĞERİ İÇERİSİNDEN ANOMALİYE SEBEP OLABİLECEK KONU BAŞLIKLARINI SEÇTİK.

df_loc['KONU'].unique()  
#'YANGIN'  'DOĞALGAZ'  'DEPREM'   'DOĞAL AFETLER'   'AFETLER'  'KAZALAR & OLAYLAR' 'İski'  'ASANSÖRLER' 'ACİL MÜDAHALE' 'SAĞLIK HİZMETLERİ' 'EMNİYET' 'GÜVENLİK'  'GÜVENLİK VE EMNİYET'  'Arıza Başvuru' 'SU'  'DSİ'  'DERE'
#'SOKAK HAYVANLARI' 'BAKIM ONARIM' 'İLÇE VETERİNER' 'MAZGAL'  'GÜRÜLTÜ'
#'TALEP' 'OTOBÜSLER' 'İETT OTOBÜSLERİ' 'METROBÜS' 'MARMARAY'  'MİNİBÜS/DOLMUŞ' 'MİNİBÜS/DOLMUŞ/SERVİS/TAKSİ' 'ÇAĞRI MERKEZİ'  'EVDE BAKIM HİZMETİ' 'SOKAKTA KALANLAR' 'HAT VE GÜZERGAH'  'SEFER GECİKMESİ/İPTALİ'  
#'RAMAZAN AYI'

df_loc['ILGILI_BIRIM'].unique()

df_loc['ILGILI_BIRIM'].value_counts().head(10)

df_loc['KANAL'].unique() 

df_loc['KANAL'].value_counts().head(10)

df_loc['ILCE'].unique()

df_loc['ILCE'].value_counts().head(10)

df_loc['ALT_KONU'].unique()

df_loc['DURUM'].unique()

df_loc['MAHALLE'].unique()

df_loc['MAHALLE'].value_counts().head(10)

df_loc

## HANDLING WITH MISSING VALUES







* MAHALLE SÜTUNUNUN YAKLAŞIK %50'LİK BİR KISMI EKSİK DEĞERLERDEN OLUŞUYORDU. BURADA SATIR BAZINDA BİR EKSİLTMEYE GİDİLDİĞİNDE 3 MİLYONLUK VERİNİN %50'LİK KISMINI KAYBETMEMEK İÇİN MAHALLE SÜTUNUNU DÜŞÜRDÜK. LOKASYON BAZINDA İHTİYACIMIZ OLAN VERİYİ İSE 'İLÇE' SÜTUNUNDAN SAĞLADIK.
* 450 ADET NAN DEĞERE SAHİP OLAN 'İLGİLİ BİRİM' SÜTUNUNDAKİ NAN DEĞERLERİ 
DÜŞÜRMEK YERİNE 'BİLİNMİYOR' DEĞERİNİ ATADIK. 
* BU SAYEDE VERİMİZDEKİ ZAMANSAL  DESENİ BOZMAMAYA ÇALIŞTIK. 

df = df_loc.drop(columns='MAHALLE', axis=1)

df.isna().sum()

df.dropna(how = 'any', axis=0, subset=['BASVURU_TANIMI'],inplace=True)

df.isna().sum()

df.dropna(how = 'any', axis=0, subset=['KONU'],inplace=True)
df.dropna(how = 'any', axis=0, subset=['ALT_KONU'],inplace=True)

df['ILGILI_BIRIM'].fillna(value='BİLİNMİYOR', inplace=True)

df.isna().sum()

df.to_csv('prepared_data.csv', index = False)


## DETAILED REVIEW OF "KONU" COLUMN




## DATA VISUALIZATION

df.reset_index(level=0, inplace=True)
df.groupby('ILCE')['BASVURU_KODU'].apply(list)

df_deneme = df.groupby('ILCE')['BASVURU_KODU'].count().head()

df_deneme = df_deneme.rename_axis('ILCE').to_frame('Başvuru_sayısı')

df_deneme

df_deneme.reset_index(inplace=True)

df_deneme.head()

df_deneme = df.groupby('ILCE')['BASVURU_KODU'].count()

df_deneme = df_deneme.rename_axis('ILCE').to_frame('Başvuru_sayısı')

df_deneme.reset_index(inplace=True)

pd.set_option('display.max_rows', None)

df_deneme

df_deneme2 = df_deneme.sort_values(by = ['Başvuru_sayısı'], ascending=False)

x = df_deneme2['ILCE']

y = df_deneme2['Başvuru_sayısı']

df["ILCE"].value_counts()
