==> Call Center Anomaly Detection/2-Data_Visualization_and_Mapping .ipynb <==
# Importing Libraries

import pandas as pd
import numpy as np
import seaborn as sns
import plotly.express as px
import datetime
import matplotlib.pyplot as plt
datetime.datetime.strptime

from sklearn.preprocessing import LabelEncoder
import matplotlib.pyplot as plt
import sqlite3
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import OneHotEncoder
import seaborn as sns

# Loading&Reading Data

df_call = pd.read_csv("/content/drive/MyDrive/ibb_data_mapping.csv")
df_loc = pd.read_csv('/content/drive/MyDrive/istanbul_ilce_koord_tr.csv', sep=';')
df_population = pd.read_excel('/content/drive/MyDrive/2019-yl-belediye-nufuslar.xlsx')
df_prepared = pd.read_csv("/content/drive/MyDrive/prepared_data.csv")

df_prepared = pd.read_csv("/content/drive/MyDrive/Colab Notebooks/prepared_data")

df_prepared

df_deneme = df_prepared.groupby('ILCE')['BASVURU_KODU'].count()
df_deneme = df_deneme.rename_axis('ILCE').to_frame('Başvuru_sayısı')
df_deneme.reset_index(inplace=True)
pd.set_option('display.max_rows', None)
df_deneme
df_deneme2 = df_deneme.sort_values(by = ['Başvuru_sayısı'], ascending=False)
x = df_deneme2['ILCE']
y = df_deneme2['Başvuru_sayısı']

plt.style.use('ggplot')

plt.figure(figsize=(20, 10))  

sns.barplot(x = df_deneme2['ILCE'], y =df_deneme2['Başvuru_sayısı']) 

plt.xticks(rotation=45)

plt.xlabel("ILCE")
plt.ylabel("BAŞVURU SAYISI")

plt.title(" ")


df_prepared['ACILIS_TARIHI'] = pd.to_datetime(df_prepared['ACILIS_TARIHI'],format='%Y/%m/%d %H:%M:%S')


df_prepared['SADE_ACILIS_TARIHI'] = df_prepared['ACILIS_TARIHI'].dt.date
df_prepared[['ACILIS_TARIHI','SADE_ACILIS_TARIHI']].head()


pd.set_option('display.max_rows', None)

df_prepared.groupby('SADE_ACILIS_TARIHI')['BASVURU_KODU'].count().head()

df_dates = df_prepared.groupby('SADE_ACILIS_TARIHI')['BASVURU_KODU'].count()

df_dates = df_dates.rename_axis('SADE_ACILIS_TARIHI').to_frame('BASVURU_KODU')


df_dates.reset_index(inplace=True)


df_dates.plot(x='SADE_ACILIS_TARIHI', y='BASVURU_KODU',figsize=(15,10))

df_anomali = df_dates.loc[252:]

df_anomali.plot(x='SADE_ACILIS_TARIHI', y='BASVURU_KODU', figsize= (15,10))


## Visualization Based on Population 

df_population.sort_values(by='2019 yılı nüfusları', ascending=False, inplace=True)

x=df_population['Belediyeler']
y=df_population['2019 yılı nüfusları']

plt.title('İlçeler Bazında Nüfus Dağılımı')
plt.xlabel('İlçe')
plt.ylabel('Nüfus')
plt.bar(x,y, color = (0.5,0.1,0.5,0.6), align='edge', )
#plt.xlabel('Saat', fontweight='bold', color = 'orange', fontsize='17', horizontalalignment='center')

## Mapping Based on Location and Calls Density

df_loc["enlem"] = df_loc["enlem"].str.replace(",",".")
df_loc["boylam"] = df_loc["boylam"].str.replace(",",".")
df_loc["enlem"] = df_loc["enlem"].astype(float)
df_loc["boylam"] = df_loc["boylam"].astype(float)

df_call.drop(columns='Unnamed: 0', inplace=True)

df = df_call.merge(df_loc, left_on="ILCE", right_on="ilce")

df.sort_values(by='TARIH',inplace=True)

df.drop(labels=[0,269],axis='index', inplace=True)
df.drop(columns="ilce", inplace=True)

df.reset_index(inplace=True)

df.drop(columns='index', axis=1, inplace=True)

df.rename(columns={'level_0':'index'}, inplace=True)

df = df.loc[2080:]

**Nisan ve Ekim ayları arasında lokasyona bağlı olarak gruplandırılmış gelen çağrı sayılarının yoğunluğunu gösteren Interaktif Mapping**

fig = px.scatter_mapbox(df, lat="enlem", lon="boylam", hover_name="ILCE", hover_data=['TARIH','GUNLUK_CAGRI_SAYISI'],animation_frame=df["TARIH"],
                         size="GUNLUK_CAGRI_SAYISI",color="GUNLUK_CAGRI_SAYISI", zoom=9.5, height=600)
fig.update_layout(mapbox_style="open-street-map")
fig.update_layout(margin={"r":0,"t":0,"l":0,"b":0})
fig.show()

from google.colab import drive
drive.mount('/content/drive')

df_lokasyon = pd.read_csv("/content/drive/MyDrive/IBB_KODLUYORUZ/sample_giz.csv")

df_lokasyon

labelencoder = LabelEncoder()

df_lokasyon['DURUM_CAT'] = labelencoder.fit_transform(df_lokasyon['DURUM'])
df_lokasyon

df_lokasyon.isnull().sum()

x = df_lokasyon["MONTH_O"]
y = df_lokasyon['diff_month']

#Çağrı açılış ayının çağrının kapanma süresine göre Bubble Plot görselleştirmesi

plt.scatter(x, y, c=x, cmap="Blues", alpha=0.4, edgecolors="grey", linewidth=2)
plt.xlabel("Açılış Ayı")
plt.ylabel("Geçen Süre")
plt.title("A colored bubble plot")
 
plt.show()

(df_lokasyon["KONU"]).groupby

df_lokasyon["KONU"].value_counts().plot.barh()

#Hangi ilçe belediyeye hangi kanallar aracılığı ile ulaştığının görselleştirilmesi

barWidth = 0.3
# Choose the height of the blue bars
bars1 = df_lokasyon["MONTH_O"]
 
# Choose the height of the cyan bars
bars2 = df_lokasyon["MONTH_C"]


height = df_lokasyon["KANAL"]
bars = df_lokasyon['ILCE']
y_pos = np.arange(len(bars))
plt.figure(figsize = (20, 15))
plt.bar(y_pos, height, color = (0.5,0.1,0.5,0.6))

plt.title('HANGİ İLÇE BELEDİYEYE NASIL ULAŞIYOR?')
plt.xlabel('İLÇELER')
plt.ylabel('KANAL')

#İlçeler ve çağrıların zamana göre ilişkisinin görselleştirilmesi

height = df_lokasyon["MONTH_O"]
bars = df_lokasyon['ILCE']
y_pos = np.arange(len(bars))
plt.figure(figsize = (20, 15))
plt.bar(y_pos, height, color = (0.5,0.1,0.5,0.6))

plt.title('İLÇELER VE BİLDİRİLEN SORUNLARIN ZAMANI İLİŞKİSİ')
plt.xlabel('İLÇELER')
plt.ylabel('AÇILIŞ AYI')

sns.regplot(x=df_lokasyon["MONTH_O"], y=df_lokasyon["MONTH_C"], marker="+", fit_reg=False)

#Ilçelere göre konuların Stack Plot ile görselleştirilmesi

x=df_lokasyon['ILCE']
y=df_lokasyon['KONU']
 
# Plot
plt.figure(figsize = (20, 15))
plt.stackplot(x,y, df_lokasyon['ILCE'])
plt.legend(loc='upper left')
plt.show()

#ilçelere göre çağrıların kapanma sürelerinin Stack Plot ile görselleştirilmesi

x=df_lokasyon['ILCE']
y=df_lokasyon['diff_days']
 
# Plot
plt.figure(figsize = (20, 15))
plt.stackplot(x,y, df_lokasyon['diff_days'])
plt.legend(loc='upper left')
plt.show()

df2 = pd.read_csv('/content/drive/MyDrive/IBB_KODLUYORUZ/new_column_csv.csv', parse_dates= ["ACILIS_TARIHI", "KAPANIS_TARIHI", "ONLY_OPEN_DATE","ONLY_CLOSE_DATE"])

# change the type of timestamp column for plotting
df2['ACILIS_TARIHI'] = pd.to_datetime(df2['ACILIS_TARIHI'])

# plot the data
df2.plot(x='ACILIS_TARIHI', y='MONTH_O')

from sklearn.preprocessing import LabelEncoder
# creating instance of labelencoder
labelencoder = LabelEncoder()
# Assigning numerical values and storing in another column
df2['KANAL'] = labelencoder.fit_transform(df2['KANAL'])
df2

from sklearn.preprocessing import LabelEncoder
# creating instance of labelencoder
labelencoder = LabelEncoder()
# Assigning numerical values and storing in another column
df2['ILCE'] = labelencoder.fit_transform(df2['ILCE'])
df2

df2.plot(x='ACILIS_TARIHI', y='ILCE')
