==> Call Center Anomaly Detection/3-FILTERING_AND_PREPARING_TO_MODEL.ipynb <==
# Importing Libraries

import pandas as pd
import numpy as np
import seaborn as sns
from datetime import datetime

# Loading&Reading Data

df = pd.read_csv("/content/drive/MyDrive/Colab Notebooks/prepared_data")

## REGEX

DATASETİNİ İNCELEDİĞİMİZ ZAMAN SOSYAL HİZMETLERE GELEN ÇOK FAZLA BİREYSEL TALEP VARDI. BUNLARI ELEMEK İÇİN, 335 KONU İÇERİSİNDEN ANOMALİYE SEBEP OLABİLECEK OLARAK BELİRLEDİĞİMİZ KONULARI DİĞERLERİNDEN AYIRMAK İÇİN BİR FİLTRELEME UYGULADIK.

df['KONU'] = df['KONU'].str.upper()
search_list = ['YANGIN','DOĞALGAZ' , 'DEPREM' , 'DOĞAL AFETLER' , 'AFETLER' , 'KAZALAR & OLAYLAR' , 'İSKİ' , 'ASANSÖRLER' , 'ACİL MÜDAHALE' , 'SAĞLIK HİZMETLERİ' , 'EMNİYET' , 'GÜVENLİK' , 'GÜVENLİK VE EMNİYET' , 'Arıza Başvuru' , 'SU' , 'DSİ' , 'DERE' , 'SOKAK HAYVANLARI' , 'BAKIM ONARIM' , 'İLÇE VETERİNER' , 'MAZGAL' , 'GÜRÜLTÜ' , 'TALEP' , 'OTOBÜSLER' , 'İETT OTOBÜSLERİ' , 'METROBÜS' , 'MARMARAY' , 'MİNİBÜS/DOLMUŞ' , 'MİNİBÜS/DOLMUŞ/SERVİS/TAKSİ' , 'ÇAĞRI MERKEZİ' , 'EVDE BAKIM HİZMETİ' , 'SOKAKTA KALANLAR' , 'HAT VE GÜZERGAH' , 'SEFER GECİKMESİ/İPTALİ']
mask = df['KONU'].str.contains('|'.join(search_list))
df_konu1= df[mask]

df_konu1

# Location Based Sample Selection

new.plot(x='ACILIS_TARIHI', y='value', figsize= (40,10))

Sarıyer lokasyonunda meydana gelen bir olağanüstü durumu görselleştirmeler sonucu tespit ettik. Bu sebeple Sarıyer ilçesini Sample olarak seçtik.

df_konu1['ILCE'] = df_konu1['ILCE'].str.upper()
search_list = ["SARIYER"]
mask = df_konu1['ILCE'].str.contains('|'.join(search_list))
selection= df_konu1[mask]
selection

selection['ACILIS_TARIHI'] = pd.to_datetime(selection['ACILIS_TARIHI'],format='%Y/%m/%d %H:%M:%S')

b=selection.value_counts().reset_index()

b.rename(columns={'0':'Value'},inplace=True)

b.sort_values(by="ACILIS_TARIHI",inplace=True)

b.reset_index(inplace=True)

c=b

c.drop(columns=['index'], inplace=True)

c.rename(columns={'0':'Value'},inplace=True)

## 15 MINUTES INTERVALS

Senaryomuza göre 15 dakikalık geriye dönük inceleme yapmak amacıyla, gelme sıklığındaki anomaliyi tespit etmek için OPEN_DATE sütununu 15 dakikalık zaman aralıklarına böldük.

frequenced_df = c.groupby(pd.Grouper(key='ACILIS_TARIHI',freq='15Min')).sum()

frequenced_df.drop(columns=["index"],inplace=True)

frequenced_df.reset_index(inplace=True)

frequenced_df.columns = ['ACILIS_TARIHI', 'value']

frequenced_df.plot(x='ACILIS_TARIHI', y='value', figsize= (40,10))

frequenced_df.to_csv('/content/drive/MyDrive/sariyer_sample.csv', index = False)
